Extract fancy-concat for use on details card.
