+++ src/routes/tools/[slug].svelte
@@ -15,56 +15,36 @@
 
 <script>
     export let tool;
+	import fancyConcat from '../../_helpers/fancy-concat';
 
-    function fancyConcat(array) {
-        let output;
-        if (array.length > 2) {
-            const lastItem = array.pop();
-            output = array.join(', ');
-            output += `, and ${lastItem}`;
-        } else {
-            output = array.join(' and ');
-        }
-
-        return output;
-    }
 </script>
 
 <svelte:head>
 	<title>{tool.name}</title>
 </svelte:head>
 
-<div class='page-header'>
+<div class="tool-detail-container">	
+	<div class="spacer"></div>
 	<h1>{tool.name}</h1>
 	<p>{tool.description}</p>
-</div>
-
-<article class="card card--full-width" data-niche={tool.niches ? Object.keys(tool.niches).join(' ') : ''}>
-	<h2>Details</h2>
-    <div class="meta">
-        {#if tool.niches }
-        <div class="card__meta"><span class="screen-reader-text">tool type: </span>
-        {#each Object.keys(tool.niches) as niche}
-            { niche }{#if tool.niches[niche].length > 0}: { tool.niches[niche].map(niche => niche.toLowerCase()).join(', ') }{/if}<br />
-        {/each}
-        </div>
-        {/if}
-        {#if tool.useCount > 0}
+	{#if tool.useCount > 0}
         <div class="card__meta">{ tool.usecount } { tool.usecount > 1 ? 'coops use' : 'coop uses' } it</div>
-        {/if}
-        {#if tool.pricing }
-        <div class="card__meta"><svg class="icon icon--pricing" viewBox="0 0 20 20" aria-hidden="true" focusable="false"><use href="/images/report.svg#report" /></svg> Pricing: { tool.pricing }</div>
-        {/if}
-        {#if tool.license }
-        <div class="card__meta"><svg class="icon icon--settings" viewBox="0 0 20 20" aria-hidden="true" focusable="false"><use href="/images/settings.svg#settings" /></svg> Licensing: { tool.license }</div>
-        {/if}
-        {#if tool.sectors && tool.sectors.length > 0 }
-        <div class="card__meta"><svg class="icon icon--sector" viewBox="0 0 24 24" aria-hidden="true" focusable="false"><use href="/images/sector.svg#sector" /></svg> Sector: { fancyConcat(tool.sectors) }</div>
-        {/if}
-        {#if tool.languages && tool.languages.length > 0 }
-        <div class="card__meta"><svg class="icon icon--info" viewBox="0 0 20 20" aria-hidden="true" focusable="false"><use href="/images/info.svg#info" /></svg> Language: { fancyConcat(tool.languages) }</div>
-        {/if}
-    </div>
-</article>
-
+    {/if}
+	<div class="spacer"></div>
+	<hr />
+	<h2>Details</h2>	
+	{#if tool.pricing }
+		<div class="card__meta"><svg class="icon icon--pricing" viewBox="0 0 20 20" aria-hidden="true" focusable="false"><use href="/images/report.svg#report" /></svg> Pricing: { tool.pricing }</div>
+	{/if}
+	{#if tool.license }
+		<div class="card__meta"><svg class="icon icon--settings" viewBox="0 0 20 20" aria-hidden="true" focusable="false"><use href="/images/settings.svg#settings" /></svg> Licensing: { tool.license }</div>
+	{/if}
+	{#if tool.sectors && tool.sectors.length > 0 }
+		<div class="card__meta"><svg class="icon icon--sector" viewBox="0 0 24 24" aria-hidden="true" focusable="false"><use href="/images/sector.svg#sector" /></svg> Sector: { fancyConcat(tool.sectors) }</div>
+	{/if}
+	{#if tool.languages && tool.languages.length > 0 }
+		<div class="card__meta"><svg class="icon icon--info" viewBox="0 0 20 20" aria-hidden="true" focusable="false"><use href="/images/info.svg#info" /></svg> Language: { fancyConcat(tool.languages) }</div>
+	{/if}
+	
+</div>
